Added in documentation for endpoint and models.
