+++ api/parent/parentRouter.js
@@ -71,7 +71,39 @@ const Parents = require('./parentModel');
  *        type: integer
  */
 
-router.get('/viz', (req, res) => {
+/**
+ * @swagger
+ * /parent/viz?childId={id}:
+ *  get:
+ *    summary: gets a Plotly line graph object for a child's squad score over time
+ *    security:
+ *      - okta: []
+ *    tags:
+ *      - Parents
+ *    parameters:
+ *      - $ref: '#/components/parameters/childId'
+ *    responses:
+ *      200:
+ *        description: A plotly graph object whose property values can be passed into plotly
+ *        content:
+ *          application/json:
+ *            schema:
+ *              type: object
+ *              properties:
+ *                data:
+ *                  type: object
+ *                  description: plug this into the "data" attribute of the plotly component
+ *                layout:
+ *                  type: object
+ *                  description: plug this into the "layout" attribute of the plotly component
+ *      401:
+ *        $ref: '#/components/responses/UnauthorizedError'
+ *      404:
+ *        $ref: '#/components/responses/NotFound'
+ *      500:
+ *        $ref: '#/components/responses/DatabaseError'
+ */
+router.get('/viz', authRequired, (req, res) => {
   const childId = req.query.childId;
 
   ops.getAll(res, Parents.getVisualizations, 'Child', childId);
