+++ lib/errorHandler.js
@@ -6,7 +6,7 @@
  * @param {String} name the singular name of the data object being operated on
  */
 const send = (res, error, name = '') => {
-  // console.log({ error: error.message });
+  console.log({ error: error.message });
   let status, message;
   if (error.message.includes('NotFound')) {
     status = 404;
