+++ lib/formatHelpers.js
@@ -140,9 +140,26 @@ const formatSubForMatchups = (pages) => {
   return res;
 };
 
+/**
+ * Takes the results of a query joining children with their submissions and parses them into
+ * the proper format for the ds api
+ * @param {Array} complexities an array of complexity scores/child name
+ * @returns {Object} returns a formatted object:\{ StudentName: '', ScoreHistory: [] }
+ */
+const formatLineGraphBody = (complexities) => {
+  const res = {};
+  complexities.forEach((score) => {
+    if (!res.StudentName) res.StudentName = score.Name;
+    if (!res.ScoreHistory) res.ScoreHistory = [];
+    res.ScoreHistory.push(score.Complexity);
+  });
+  return res;
+};
+
 module.exports = {
   formatCohortSubmissions,
   formatProfiles,
   formatTeam,
   formatSubForMatchups,
+  formatLineGraphBody,
 };
